Performed alot of changes as mentioned in tutorial 2
Added timezone into the settings file. Created Question and Choice
models. Added polls in the INSTALLED_APPS in settings.py. Ran
makemigrations to and then ran the migrate command to make the models in
the database. Made the poll app visible and modifiable in the admin page
by registering it in the polls/admin.py file