Tutorial 7- Customizing admin page for polls application.
Modified the admin.py file to make the admin site more user friendly and
more efficient to use.
Also modified models.py file for the same purpose.
Added a template folder and used the django base_site template and
modified it to change the title of the admin page to polls
administration.