Additional feature added: Allows users to add their own questions to the site
*Now not only the admin, but users too are able to add questions and
choices to the site. This will allow users to add questions without
giving them administrator status or without any administrator
intervention. (only drawback in this version - each user-entered
question must have exactly 4 choices.)
Changes to files-
forms.py contains the structure of the form. It is used along with the
name.html template to display the form.
urls.py file is updated to point to the views.get_name function which
allows users to submit their questions.
Added a link in the index.html template for adding user's questions.